Raise upper constraint for PasteDeploy
I've recently released a new version of PasteDeploy, the main change
in it is the removal of a warning produced when running under
Python3 (a source of much noise in some functional tests).

The major version update is the result of a discussion on how best
to signal that the project now lives under the Pylons project and
has moved from BitBucket to GitHub[1]. It's supposed to be backwards
compatible.

[1] https://github.com/Pylons/pastedeploy

Change-Id: I6132a21eadc345b41c6b3c95a9e40fb1fd5e0c22