Cap grpcio to below 1.16.0
grpcio-1.16.0 introduced a bug that is breaking the tooz unit tests.
We can't use any higher versions of grpcio until that is fixed.

Is the library actively maintained? Yes

Is the library good code? Yes

Is the library python 3 compatible? Yes

Is the library license compatible? Yes, Apache 2.0

Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)? It's a transitive dependency, so it must be.

Is the function of this library already covered by other libraries
in global-requirements.txt? No

Is the library required for OpenStack project or related dev or
infrastructure setup? Yes, tooz

Change-Id: I8b026a0680a91a68281e354a59012565f4db1b95
Related-Bug: 1808046