Add octavia-lib to global-requirements
octavia-lib is a spin out of code from the octavia project that is
required to support third-party provider drivers.

Is the library actively maintained?
Yes
Is the library good code?
Yes
Is the library python 3 compatible?
Yes, it has py3 gates
Is the library license compatible?
Yes
Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
Not yet
Is the function of this library already covered by other libraries in global-requirements.txt?
No
Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes, Octavia
If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?
Yes
Do I need to update anything else?
No, this is existing code pulled out of the octavia project

Change-Id: Ia9e067b2bebc2ea8732fbc5320887c3df4cbbf9c