Add os-resource-classes to global requirements
As version 0.1.0 is the only release (thus far), also add it
manually to upper-constraints.

os-resource-classes is a library of resource class symbols used by
the placement service and clients thereof, created to allow multiple
packages to reliably use the same symbols, similar to the os-traits
library.

It is managed by the placement portion of the nova project and is
licensed with apache 2.0.

Change-Id: Iff4b1f3bb025ecbd7066844b8f29e10e892b45bb