Increase max version of sqlalchemy-migrate to 0.12
sqlalchemy-migrate 0.12 included an important fix for newer versions of
SQLite. Without this fix, testing older stable branches of nova and
other projects using the package is no longer possible on Fedora.
Unfortunately, sqlalchemy-migrate does not use stable branches, but the
change set between 0.11 and 0.12 is minimal enough, consisting of some
small non-functional changes (spellings, gitignore file additions),
removal of an upper limit on pbr version, and updates to resolve
deprecation warnings that were present on the older branches too. As
such, we can simply enable the use of 0.12 on the older stable branches,
resolving the issue without the need to introduce stable branches or
anything like that.

Change-Id: I55077cea917f630cd84c3dc83fd7e79230abcf24
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>