Add protobuf to global-requirements
This commit adds protobuf into global-requirements
to make it possible to use PodResources service
in kuryr-kubernetes project.

Is the library actively maintained?
Yes - https://github.com/protocolbuffers/protobuf

Is the library good code?
Yes

Is the library python 3 compatible?
Yes

Is the library license compatible?
Yes - BSD License (3 clause)

Is the library already packaged in the distros we
target (Ubuntu latest / Fedora latest)?
As far as I know, Ubuntu latest includes

Is the function of this library already covered by
other libraries in global-requirements.txt?
No

Is the library required for OpenStack project or related
dev or infrastructure setup? (Answer to this should be Yes,
of course) Which?
Yes, kuryr-kubernetes

If the library release is managed by the Openstack release
process does it use the cycle-with-intermediary release type?
No

Do I need to update anything else?
No

Change-Id: I2c6ace4f2dffa6a3f6d53eb2fdce71efc2883b85
Signed-off-by: Danil Golov <d.golov@samsung.com>