Bump pylxd to 2.2.7
We cannot build wheels for pylxd 2.2.5 anymore. Its 2.2.5 setup.py
file install_requires have no upper bound on requests, and wheel
cannot be built with latest requests.

Bumping to 2.2.7 is bumping to the lowest version which contains
a patch separating the install and setup_requires [1], allowing
us to build its wheels. Anything above the Queens branch is not
impacted by the issue.

[1]: https://github.com/lxc/pylxd/commit/4b8ab1802f9aee4eb29cf7b119dae0aa47150779

Change-Id: I97b909385a8e6778a1a09afb26b8c33be1f294ca