Add upper-constraints enforcement for local reqs
This adds the use of upper constraints in our tox environments. Jobs
like openstack-tox-py27 have started failing due to trying to install
lib versions that only support Python 3.

The 'generate' target keeps the unconstrained deps since we actually do
need the latest installed for that.

Updates the requirements-integration job to run using py36 since that
now fails with running under py27. This should get the job to pass until
we are able to rewrite it to not be a legacy job.

NOTE(dmllr): updated install_command to include local constraints file

Change-Id: I2598503c2d4397d7efbb03a76df9a4e91dd0c1cf
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>