Set upper-constraint for scikit-learn
We capped scikit-learn for Python 2 compatibility in change
If558f184c959e4b63b56dec3ca1571d1034cfe5c, but upper-constraints were
not set. When scikit-learn is a dependency of a dependency, like with
taskflow, this can result in consuming projects still getting the newer
version under py2 runtimes. This adds the upper-constraint so the cap
will be enforced.

Change-Id: If205849b1248c5295d087b747c9703d0c317810e
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>