Fix where we use constraints
in I2598503c2d4397d7efbb03a76df9a4e91dd0c1cf we mistakenly added
constraints to install_command in out based testenv.  This is incorrect
as it means the generate env will be run in a constrained environment.

We don't use UPPER_CONSTRAINT_FILE as *right now* it's the same thing
and this leaves us internally consistent (even if technically incorrect)

Change-Id: Ia61d944dbfdf494598e41e5fbefd0e32a91814ed