Add tripleo-ansible to the requirements sync
Answers to the basic requirements questions.

Is the library actively maintained?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the library good code?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the library python 3 compatible?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the library license compatible?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the function of this library already covered by other libraries in global-requirements.txt?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Is the library required for OpenStack project or related dev or infrastructure
setup? (Answer to this should be Yes, of course) Which?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

If the library release is managed by the Openstack release process does it use
the cycle-with-intermediary release type?

N/A - TripleO-Ansible is not a library and not bringing in new libs within
OpenStack.

Do I need to update anything else?

no.

Change-Id: I675dfc70d9237fa1c4873a06d6b9bad0bbc1b8e0
Signed-off-by: Kevin Carter <kecarter@redhat.com>