Remove requirements-integration job
This job was a legacy job that attempted to verify each project could be
independently installed after all global-requirements had been
installed.

This job is no longer necessary since we have much better coverage with
requirements-check and other jobs that make sure the requirements are
compatible. It is also a legacy zuulv2 job, uses the long deprecated
devstack-gate, and runs with Python 2.7.

For historical reference, this was added originally to PBR back in 2013
[0] and moved to requirements in 2014 [1].

[0] https://opendev.org/openstack/pbr/commit/378261a8cba818a29b7e3d33b917de5ed0751579
[1] https://opendev.org/openstack/requirements/commit/8e073738db06086f949d2be377999356aee2d8d0

Change-Id: I29abcba929995f9307c392110160d889a29b47c3
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>