Remove most python-2.7 jobs
remove all jobs depending on projects that are set to remove support
for python-2.7 during the Ussuri cycle.  This will keep us from being
surprised when a project pulls support for py27 and breaking our gate.

Originally this was set to occur later in the cycle but as we depend
on the projects for cross jobs it needs to be done earlier.

Change-Id: I6c9f545e09779431120cee942318013ed91a581a
Signed-off-by: Matthew Thode <mthode@mthode.org>