Add horizon to upper-constraints
horizon plugins use horizon as a library.
Having horizon in upper-constraints would be helpful
to avoid unexpected breakage due to a newer horizon releases,
particularly in stable branches.

The reason we don't have horizon in upper-constraints so far is
because most horizon plugins would like to work and test with horizon
in a same branch, but Zuul v3 allows such tests.
This reasoning is no longer true.

Updated to stable/rocky maximum of 14.0.4.

Note that this is proposed to stable branches directly.
In the master branch, it is required to split out common portions
in horizon into a separate deliverable and it was rejected to
add horizon to upper-constraints.txt.

Change-Id: I421ee2d498ea5100d9b52e356b46f2d594638b5b
(cherry picked from commit 3e54878f9c87266e1b9775f72af06de27a3dad65)