Cap Django again
I propose to cap Django version again based on the following considerations.

Django minor bumps usually drop deprecated features.
According to the Django document [1], it is an expected change.
Django defines the consistent policy and I think it is not easy
to change the versioning policy in near future. (It would take time
even if it happens.)

If horizon is the only project to consume Django, we could probably
catch up latest Django versions, but we have many horizon plugins.
It is not realistic to update all horizon plugins promptly.
As a fact, Django 2.2 adoption in all horizon plugins was quite
straighforward but it took over a month.
Django 2.0 adoption took over three months.

On the other hand, we would like to consume the latest patch version of
Django, i.e. 2.2.x for Django 2.2 series. If we use blacklist.txt,
we will never get a newer version without manual process.

Considering these, I believe capping Django makes sense.
Django minor version is used for capping as minor versions can
bring backward-incompatible changes.

[1] https://www.djangoproject.com/download/#supported-versions

FYI:
horizon and plugins have jobs with Django versions different from
the Django version in upper-constraints. The steps to bump the upper
bound of Django would be:
(0) Ensure horizon and plugins to work with a newer Django version
    (using non-primary Django version jobs)
(1) Bump the upper bound of Django in global-requirements.txt
(2) Sync requirements.txt in horizon with global-requirements
(3) Cut a new release of horizon so that plugins can consume
    a newer Django version
(4) Bump Django version in upper-constraints.txt

Change-Id: Ie51b8719abc3efb3e4fa2953ba3596ed1b214418