Add pywinrm to global-requirements.txt
pywinrm allows sending commands to remote Windows hosts via WinRM,
pretty much resembling SSH.

- Is the library actively maintained?

Yes, there are a few quite recent releases:
https://github.com/diyan/pywinrm/releases

- Is the library good code?

Yes. Despite being pre-1.0, it provides a stable and well documented API. The
code is easy to understand and well tested. It's a simple library that
implements the WinRM SOAP protocol.

- Is the library python 3 compatible?

Yes.

- Is the library license compatible?

Yes, it's MIT licensed.

- Is the library already packaged in the distros we target (Ubuntu latest /
Fedora latest)?

Yes, it's available on PyPI as well as Ubuntu upstream repositories.

- Is the function of this library already covered by other libraries in
global-requirements.txt?

No.

- Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes, it's already used by the Manila Windows driver (similar to the Linux
generic one but using Windows instances) as well as the oswin-tempest-plugin
project.

- If the library release is managed by the Openstack release process does
it use the cycle-with-intermediary release type?

No, it's not an openstack project.

- Do I need to update anything else?

No.

Change-Id: I57f958ca93a2dd424ce988cd28816550bbd6e2d9
Closes-Bug: #1745823