Add importlib-metadata
This Apache-2.0 library is intended to end up in python 3.8's stdlib.
For the rest of us we can get access to it via pypi on python2.7 and
python>=3.5. Why is this important? This package replaces pkg_resources
entrypoint and metadata APIs which are inefficient. This will give us a
head start in using the new more efficient tools for entrypoint
management.

Needed-By: https://review.opendev.org/#/c/661856/
Change-Id: Icd79bd057b9578a9ec8b107cc5d1ac926899c34b