Add Pygments to upper-constraints
Pygments is already in global-requirements, but not in
upper-constraints. Pygments is required by Sphinx.

Pygments removed support for py2 in version 2.6.0 so we
need to ensure that it is properly pinned to prevent
tests/docs failing.

Pygments was added to the blacklist in
I643c7632a02c87820c9c51f03fa34f0764384783 but it's not
entirely clear why. To ensure that it is managed in
upper-constraints from now on, we remove it from the
blacklist.

Change-Id: I18a50482e588176047ede81c4ee994c9d7df282b
Closes-Bug: #1869174