Add validations-libs and validations-common to openstack_requirements
validations-libs and validations-common are needed to the
openstack_requirements

Answers to the basic requirements questions.

Is the library actively maintained?

Yes, the libraries is actively maintained and mainly used by TripleO projects

Is the library good code?

Yes the libraries is covered by CI and Unit tests

Is the library python 3 compatible?

Yes

Is the library license compatible?

Yes

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Yes

Is the function of this library already covered by other libraries in global-requirements.txt?

No

Is the library required for OpenStack project or related dev or infrastructure
setup? (Answer to this should be Yes, of course) Which?

Yes, it's required for python-tripleoclient and TripleO projects

If the library release is managed by the Openstack release process does it use
the cycle-with-intermediary release type?

Yes

Do I need to update anything else?

No.

Change-Id: Idcf64b4c83f9004bb326f66c8c0da80bb8817a2c