Add back dfs-sdk
This package was removed in cleanup patch
I158a8c449d6da6d9e45eba645cf0fd04320a972e, but it is defined as an extra
in the openstack/cinder setup.cfg file. That file must have been skipped
in the check for its use.

Based on the script that has been added to check for unused packages,
this shouldn't happen again. For now, even though this is an optional
dependency that is only installed when requested, it needs to be added
back to prevent the requirements check job from failing other updates to
the cinder repo requirements.

Change-Id: I2fd1ecae8834169f29196c5762518cb43c06e496
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>