Add validations-libs to the global-requirements
tripleoclient needs the validations-libs as dependencie, we need to
add the validations-libs to the global-requirements:
https://review.opendev.org/#/c/714629/

Answers to the basic requirements questions.

Is the library actively maintained?

Yes, the libraries is actively maintained and mainly used by TripleO projects

Is the library good code?

Yes the libraries is covered by CI and Unit tests

Is the library python 3 compatible?

Yes

Is the library license compatible?

Yes

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Yes

Is the function of this library already covered by other libraries in global-requirements.txt?

No

Is the library required for OpenStack project or related dev or infrastructure
setup? (Answer to this should be Yes, of course) Which?

Yes, it's required for python-tripleoclient and TripleO projects

If the library release is managed by the Openstack release process does it use
the cycle-with-intermediary release type?

Yes

Do I need to update anything else?

No.

Change-Id: Iec6b639fce37135f8840240b5d2a1a168bf9bb4e
(cherry picked from commit 5368475655b129ed71a6c207220570a36a93b3de)