Add ansible to blacklist
Some projects, such as kolla-ansible, have a dependency on ansible. At
present it is not possible to add this to your requirements as it does
not appear in the global requirements file. However, it is useful to
have ansible as dependency to allow testing of plugins.

There is not a strong dependence on the openstack libraries, so it makes
sense to include this in the blacklist and not in global requirements,
so that the individual projects can specify the version required.

The reason for Ussuri backport: Kolla cut off the Ussuri branch
already in Victoria cycle, with this in place, and now it fails
on req-checks jobs in Ussuri.

Change-Id: Ie30fe79e686f89fbdb246b5330d5b9f10aecf354
(cherry picked from commit dcef447f6d5233ecc20ba19b7fa1e04e7dfb740c)