Add yamllint to blacklist
Some projects, such as kolla-ansible, use yamllint for linting YAML
code.
At present it is not possible to add this to your requirements as it
does not appear in the global requirements file.

The reason for Ussuri backport: Kolla cut off the Ussuri branch
already in Victoria cycle, with this in place, and now it fails
on req-checks jobs in Ussuri.

Change-Id: Ibf26d927cd45d0f39a0fcb7c4700786dbf002459
(cherry picked from commit 6ca51f2b1c52a6691925029586b6514ede8f13a1)