Add pytest-xdist to global requirements
pytest was introduced in [1] as horizon requirements. Horizon
wants to speedup tests execution time now, so we need this
plugin to be included into requirements.

Is the library actively maintained?
According to GitHub [2] project has active community with
regular releases.

Is the library good code?
It has a good number of unit tests and uses black for code
formatting.

License: MIT

Is the library already packaged in the distros we target?
Yes, there are packages in Ubuntu and Fedora repositories.

Is the function of this library already covered by other
libraries in global-requirements.txt?
No.

Is the library required for OpenStack project or related
dev or infrastructure setup?
It's used only for tests.

[1] Idba192945ee8980b6ee2d49ce9651dbedb479609
[2] https://github.com/pytest-dev/pytest-xdist

Change-Id: I5e0e8e6a5af0f439a0b3e972d9a9ed07fcaaddcd