Add more wheel builds to check
This adds centos-8 and ubuntu-focal to the x86 wheel builds.

It also adds the arm64 builds in the check-arm64 queue.  This is
currently a non-voting queue -- however failures here will actually
stop the periodic wheel release job.  There might be an argument for
moving them into check/gate; we have been a little worried that
insufficient arm64 resources would hold up the gate for changes that
required timely merge; however the linaro cloud is certainly seeming
to be stable and performant.  These do not work right now, and are
fixed up in follow-ons.  We can start here as we initially debug.

Change-Id: I9ae13fdb702dd2f140bebd7db326b3b47ef7f284