Ensure pip is installed locally
Recent changes to the base nodepool images removed the system-wide
installation of pip and virtualenv. As a result of this, some
requirements-check jobs are failing with an error that pip could not be
found.

This adds the ensure-pip role as part of pre_tasks in our
requirements-check playbook.

Change-Id: Id4969156383545eb94462677e3321b89c1160b48
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit 74a5bed896893ecd3bdedb100439f773af7c7f5e)