Switch node tests to inherit nodejs-run-test-browser.
Current base of nodjs-npm-run-test is deprecated. We need to switch to
the newer base job to pick up new behavior. Horizon npm tests require
a browser so the job inherit nodejs-run-test-browser now.

In addition, tox is not anymore pre-installed on the OpenDev CI images.
npm jobs in horizon need tox, so we need to install it in pre-run phase.

Change-Id: Iffa3f7631bec0a5f7b0c701df87305725cae5708
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>