Temporarily cap setuptools<48.0.0
with new setuptools 48.0.0, devstack started failing on
 - usr/local/bin/keystone-manage: No such file or directory"

https://zuul.opendev.org/t/openstack/build/55e53d1c37194d8992f79a4090da86ee/log/job-output.txt#10572

It is blocking all the gate for now. We are not sure if that is bug on
setuptools side or it is incompatible change and we need to update devstack
for that but till we figure out that let's use the old setuptools to unblock the gate.

Change-Id: Ic9ee4d5164b144638c3ea3d4ddc1e6bac86378c0
ML: http://lists.openstack.org/pipermail/openstack-discuss/2020-July/015779.html
Related-Bug: #1886237