Add rbd-iscsi-client to global-reqs
This patch adds the new rbd-iscsi-client to global-requirements.
This client lib is used for cinder ceph iscsi driver.

Currently, the library is maintained by me, but have plans to
possible add it under cinder's umbrella.

The library is python3 compatible.

Apaache 2 license.

The library is currently not packaged by a distro yet as far
as I know.

The function of this library is not already covered by other
libraries.

The library is required by the new cinder ceph iscsi driver.
https://review.opendev.org/#/c/662829/

Change-Id: I6580092f6f61e60e743fef18bc6e6ded08b97ae3