Update rbd-iscsi-client to 0.1.8
This patch updates the upper-constraints.txt to
accomodate the new version of rbd-iscsi-lient for
the cinder ceph-iscsi driver.

This has a new required parameter to delete_disk
that allows the preservation of the backing rbd image,
to ensure cleanup of the iscsi gateway on nova detach
without deleting the rbd image from the pool.

Change-Id: I980190c4f9ce30cc01aa4c353ea976249c1112d2