Reintroduce setuptools <50 cap
Per the linked github issue, the setuptools has re-introduced it's
vendoring of distutils so it not longer obeys the overrides configured
by Debuntu.  This was last pinned with
ca1eb8dd0ae9c7c7ac149e8b4ff9bbc9630eb43f; setuptools reverted their
change, but it is back now.

In short, Debuntu overrides distutils to install everything in
"dist-packages", leaving "site-packages" in case you install a
non-packaged interpreter which has that all to itself (i.e. packaged
Python doesn't mix/conflict with a non-packaged Python install).

The latest setuptools now ignores the overrides Debuntu ships in
distutils, and thus installs binaries to /usr/bin and libraries to
site-packages; the first causing problems for numerous scripts which
assume /usr/local/bin/<foo> and the second meaning that
/usr/bin/python3 (i.e. Debuntu packaged python) can't find anything
installed anyway (because it's in site-packages, which isn't
searched).

This pin affects the devstack pip install from [1].

At this stage, we do not have a solution.

[1] https://opendev.org/openstack/devstack/src/commit/e7625fc72cbbbfaf37a6682166bc2180d4e50855/tools/install_pip.sh#L153
[2] http://lists.openstack.org/pipermail/openstack-discuss/2020-August/016905.html

Change-Id: I97e8fe6dd296d8552e256dc1dd85e9d13a9feafe