Move ansible-runner to 1.4.0 to fix unclosed fd in 1.3.4
In ansible-runner 1.3.4 process isolation cleanup was failing
because some file descriptor for the stdout in the runner
was not closed.
The close fd function was a noop (pass), so a fix has been
merged and pushed in 1.4.0 in order to cleanly flush and
close the fd.

This commit ask to bump to this version in order to get
a clean release of ansible-runner in stable/train.

Change-Id: I192f8ce533c2668facae0ed44834097282ef3be2