Revert updating version of pymemcache
The latest version of pymemcache (3.4.0) breaks the tooz memcache
driver effectively breaking all the projects that rely on that.
An example is provided by the ironic-inspector-tempest-managed-non-standalone
job in the ironic-inspector project where we had to force the
pymemcache version to be installed to 3.3.0
With this patch we set the version of pymemcache in upper-constraints
back to 3.3.0 and we blacklist version 3.4.0 in global-requirements
to prevent the use globally.

Change-Id: I155024516774949d9438d4f0b918f023eebcc6fc