Remove cap on PrettyTable
PrettyTable was capped at a < 0.8, which meant we were getting the
veritably ancient 0.7.2 release first release in April 2013 (!) [1].
The project is now being maintained as a Jazzband project [2], meaning
we can remove the cap. The API is pretty much unchanged so this should
have little impact for projects that depend on this library.

[1] https://pypi.org/project/prettytable/#history
[2] https://github.com/jazzband/prettytable

Change-Id: I42d2fe96cd81d8c159f54dc10a18bf6bfebf601f
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>