Add neutron to stable upper-constraints
Some projects, notably neutron plugins, are using the neutron service as
a library. There is the neutron-lib package, but progress is slow and
everything need was not moved there yet.

This is being added directly in stable branches only. Master needs to
get the latest version available, so these will need to be added at each
release so the stable branches use the correct version of neutron for
their respective stable release.

To allower upper capping, this commit also drops neutron from blacklist

Change-Id: I74d378d20287d78c57e968d744222da345c85596