Add ruamel.yaml to global-requirements.txt
This Python library is now a requirement for building
documentation (ansible-autodoc) for the following project:
- openstack/tripleo-validations
- openstack/validations-common

ruamel.yaml is a YAML 1.2 loader/dumper package for Python.

The library is actively maintained, with almost a weekly release. It is
compatible with Python 2 and 3 and licensed under the MIT.

https://pypi.org/project/ruamel.yaml/

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I208b00629493953a502f8b95756820ae0d9a9cb6
(cherry picked from commit 5930f2fbd44667943b9ea20fcaf995f25dde4954)