Add datetimerange library to global requirements
While developing a new feature in CloudKitty, we adopt a library
(datetimerange) to facilitate the handling/processing of
datetime ranges. Therefore, to make the Zuul requirements check
happy, we are introducing this new dependency into the global
requirements file.

Add datetimerange library to global requirements

While developing a new feature in CloudKitty, we adopt a library
(datetimerange) to facilitate the handling/processing of
datetime ranges. Therefore, to make the Zuul requirements check
happy, we are introducing this new dependency into the global
requirements file.

Here follows the answers to the new library requirement questions:

1 - Is the library actively maintained?

Yes, it is. As we see in [1], the component has more or less one
yearly release. It is a very focused library, so this release process
makes sense.

2 - Is the library good code?

Yes, it is. As we can see in [2], it has a 95% of code coverage.
Moreover, while navigating in the code repository, we can see that
methods are well documented, and coded.

3 - Is the library python 3 compatible?

Yes, it is. That is also found in the README of the project in [2].

4 - Is the library license compatible?

According to reference [3]:

```
In order to be acceptable as dependencies of OpenStack projects,
external libraries (produced and published by 3rd-party developers)
must be licensed under an OSI-approved license that does not restrict
distribution of the consuming project. The list of acceptable licenses
includes ASLv2, BSD (both forms), MIT, PSF, LGPL, ISC, and MPL.
Licenses considered incompatible with this requirement include GPLv2,
GPLv3, and AGPL.
```

Therefore, the project fits nicely into these requirements. The project
in question is licensed under MIT, as we can see in [2], in the right
corner part of the page, and also in the LICENSE file at the root of
the project.

5 - Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

It is installed via PIP; PIP is available in major distros. Therefore,
it is possible to use this library in all of them by issues the command
`pip install DateTimeRange`.

6 - Is the function of this library already covered by other libraries
in global-requirements.txt?

No. At least, I did not find it.

7 - Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes, it is. We have a new feature being developed ion Cloudkitty in
[4], which needs this library to handle DateTime ranges.

8 - If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?

It is not managed by the OpenStack release process.

9 - Do I need to update anything else?

No.

[1] https://github.com/thombashi/DateTimeRange/releases
[2] https://github.com/thombashi/DateTimeRange
[3] https://governance.openstack.org/tc/reference/licensing.html
[4] https://review.opendev.org/c/openstack/cloudkitty/+/799207

Change-Id: I608331e28ef349399068f9b1fa16bd6a6c5779e6