Add autopage to global-requirements
autopage is a library for automatically invoking 'less' when needed in
command-line programs and piping the output through it. It is wanted to
improve the user experience of python-openstackclient and other
cliff-based command line tools.

Review questions:
* Is the library actively maintained?

Yes (by me).

* Is the library good code?

I wrote it, so I'd like to think so. I've been testing it out in one of
my own projects in a lot of different scenarios, and I believe it has
got to the point where it is very robust.

* Is the library python 3 compatible?

Yes (requires Python 3.6 or later).

* Is the library license compatible?

Yes, Apache 2.0 license.

* Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?

No, but the necessary packaging files are available upstream, and distro
packages for Fedora/CentOS and Ubuntu are available in Copr and PPA,
respectively:

https://copr.fedorainfracloud.org/coprs/zaneb/autopage/packages/
https://launchpad.net/~zaneb/+archive/ubuntu/autopage

* Is the function of this library already covered by other libraries in
  global-requirements.txt?

Not that I know of.

* Is the library required for OpenStack project or related dev or
  infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes, of course.

https://review.opendev.org/c/openstack/cliff/+/799344
https://review.opendev.org/c/openstack/cliff/+/799345

* If the library release is managed by the Openstack release process
  does it use the cycle-with-intermediary release type?

N/A

* Do I need to update anything else?

No, there are no transitive dependencies.

Change-Id: Ic773519e5ab3e940cb23a96b99295f7eb8986691